Listener framework for asynchronous search (#2)
Summary
Contains the basic framework needed to support search wait_for_completion timeouts. The listener supports non-blocking wait for timeouts constructs, allows the clients to wait till the timeout or a search completion, whichever finishes earlier. The listener is guaranteed to be invoked exactly once. Each asynchronous search can have multiple listeners attached, which should detach the moment the timeout expires. All listeners are eventually removed once the final search response arrives. Any concurrent listener attach would either be added to the list of listeners to be invoked or be invoked immediately basis the completion of the search on a separate thread.