Perform a full reduce on incremental partial reduces on demand (#10)
The partial reduce phase does an incremental however the buckets keep accumulating. When a GET request is made(on demand) for a partial asynchronous search response, a final reduce needs to be performed to reduce to the desired bucket. Too many buckets can create high memory pressure espl when the results aren't yet finalised.